scholarships refactor
