+++ apps/admin-dashboard/app/routes/_dashboard.students.import.scholarships.tsx
@@ -9,15 +9,9 @@ import {
   redirect,
 } from '@remix-run/node';
 import { Form as RemixForm, useActionData } from '@remix-run/react';
-import dayjs from 'dayjs';
 import { z } from 'zod';
 
-import { db } from '@oyster/db';
-import {
-  Email,
-  ScholarshipRecipient,
-  type ScholarshipType,
-} from '@oyster/types';
+import { importScholarshipRecipients } from '@oyster/core/scholarships';
 import {
   Button,
   FileUploader,
@@ -26,11 +20,8 @@ import {
   Modal,
   validateForm,
 } from '@oyster/ui';
-import { id } from '@oyster/utils';
 
-import { parseCsv } from '@/admin-dashboard.server';
 import { Route } from '@/shared/constants';
-import { findStudentByEmail } from '@/shared/queries/student';
 import {
   commitSession,
   ensureUserAuthenticated,
@@ -89,97 +80,6 @@ export async function action({ request }: ActionFunctionArgs) {
       'Set-Cookie': await commitSession(session),
     },
   });
-}
-
-const TYPE_FROM_CSV: Record<string, ScholarshipType> = {
-  Conference: 'conference',
-  Direct: 'direct',
-  Tuition: 'tuition',
-};
-
-const ScholarshipRecipientRecord = z.object({
-  Amount: z.coerce.number(),
-
-  // If the date is valid, then we'll just set to the 12th hour of the day. We
-  // chose that arbitrarily so that all timezones would show the same date.
-  'Award Date': z
-    .string()
-    .refine((value) => dayjs(value).isValid())
-    .transform((value) => dayjs(value).hour(12).toDate()),
-
-  Email: Email,
-
-  Reason: z.string(),
-
-  // In order for the CSV to be more friendly to our admins, we'll require
-  // them to use the title-case version of the scholarship type.
-  Type: z.string().transform((value) => TYPE_FROM_CSV[value]),
-});
-
-async function importScholarshipRecipients(
-  input: ImportScholarshipRecipientsInput
-) {
-  const csvString = await input.file.text();
-
-  const records = await parseCsv(csvString);
-
-  if (!records.length) {
-    throw new Error(
-      'There must be at least one row in order to import scholarship recipients.'
-    );
-  }
-
-  const scholarshipRecipients = await Promise.all(
-    records.map(async (record, i) => {
-      const result = ScholarshipRecipientRecord.safeParse(record);
-
-      if (!result.success) {
-        throw new Error(
-          `There was an error parsing row #${i} (${record.Email}).`
-        );
-      }
-
-      const {
-        'Award Date': awardedAt,
-        Amount: amount,
-        Email: email,
-        Reason: reason,
-        Type: type,
-      } = result.data;
-
-      const student = await findStudentByEmail(email)
-        .select(['students.id'])
-        .executeTakeFirst();
-
-      if (!student) {
-        throw new Error(`Could not find student with the email "${email}".`);
-      }
-
-      return ScholarshipRecipient.omit({
-        createdAt: true,
-        deletedAt: true,
-        updatedAt: true,
-      }).parse({
-        amount,
-        awardedAt,
-        email,
-        id: id(),
-        reason,
-        studentId: student.id,
-        type,
-      });
-    })
-  );
-
-  await db
-    .insertInto('scholarshipRecipients')
-    .values(scholarshipRecipients)
-    .onConflict((oc) => oc.doNothing())
-    .execute();
-
-  return {
-    count: scholarshipRecipients.length,
-  };
 }
 
 export default function ImportScholarshipsPage() {
