+++ packages/core/package.json
@@ -30,6 +30,7 @@
     "./resume-books": "./src/modules/resume-book/resume-book.core.ts",
     "./resume-books.types": "./src/modules/resume-book/resume-book.types.ts",
     "./resume-books.ui": "./src/modules/resume-book/resume-book.ui.tsx",
+    "./scholarships": "./src/modules/scholarship/scholarship.core.ts",
     "./slack.server": "./src/modules/slack/index.server.ts"
   },
   "scripts": {
