+++ packages/types/src/index.ts
@@ -23,7 +23,6 @@ export type { GetActivityType } from './domain/gamification';
 export { ProfileView } from './domain/profile-view';
 export { Program, ProgramParticipant } from './domain/program';
 export { Resource, ResourceStatus, ResourceUser } from './domain/resource';
-export { ScholarshipRecipient, ScholarshipType } from './domain/scholarship';
 export {
   ActivationRequirement,
   MemberEthnicity,
